adds google analytics v4
