+++ .env.development
@@ -8,3 +8,4 @@ NEXT_PUBLIC_CALLBACK_URL=http://localhost:$PORT/auth
 NEXT_PUBLIC_BLOG_API_URL=https://blog.resourcewatch.org/wp-json/wp/v2
 NEXT_PUBLIC_WRI_API_URL=https://api.resourcewatch.org
 NEXT_PUBLIC_CONTROL_TOWER_URL=https://api.resourcewatch.org
+NEXT_PUBLIC_GOOGLE_ANALYTICS_V4_ID=G-PTF4BE2G4G
