+++ .env.production
@@ -9,3 +9,4 @@ NEXT_PUBLIC_RW_ENV=production
 NEXT_PUBLIC_APPLICATIONS=rw
 NEXT_PUBLIC_BLOG_API_URL=https://blog.resourcewatch.org/wp-json/wp/v2
 NEXT_PUBLIC_WRI_API_URL=https://api.resourcewatch.org
+NEXT_PUBLIC_GOOGLE_ANALYTICS_V4_ID=G-PTF4BE2G4G
