+++ layout/head/app/index.js
@@ -5,7 +5,6 @@ import HeadApp from './component';
 
 export default connect(
   (state) => ({
-    routes: state.routes,
     hostname: state.common.hostname,
   }),
   null,
