+++ scripts/userreport.jsx
@@ -1,0 +1,28 @@
+import React from 'react';
+import Head from 'next/head';
+
+export default function UserReportScript() {
+  return (
+    <Head>
+      <script
+        type="text/javascript"
+        // eslint-disable-next-line react/no-danger
+        dangerouslySetInnerHTML={{
+          __html: `
+            window._urq = window._urq || [];
+            _urq.push(['setGACode', '${process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS}']);
+            _urq.push(['initSite', '085d5a65-977b-4c3d-af9f-d0a3624e276f']);
+            (function() {
+            var ur = document.createElement('script');
+            ur.type = 'text/javascript';
+            ur.async = true;
+            ur.src = ('https:' == document.location.protocol ? 'https://cdn.userreport.com/userreport.js' : 'http://cdn.userreport.com/userreport.js');
+            var s = document.getElementsByTagName('script')[0];
+            s.parentNode.insertBefore(ur, s);
+            })();
+          `,
+        }}
+      />
+    </Head>
+  );
+}
